requirements.txt
feedparser>=6.0.0
anthropic>=0.40.0
pyyaml>=6.0
aiohttp>=3.9.0
jinja2>=3.1.0
python-dateutil>=2.8.0
